packing lambda_package for deploying
