+++ elevenlabs-poc/.example.env
@@ -1,0 +1,4 @@
+ELEVENLABS_API_KEY=
+
+AWS_ACCESS_KEY_ID=
+AWS_SECRET_KEY=
