+++ elevenlabs-poc/README.md
@@ -1,1 +1,8 @@
-README
+## Packing zip file to deploy to AWS Lambda
+* Make sure folder `src` is latest updated.
+* Run command:
+```
+make packing
+
+```
+* The package should be saved in `build` folder. Upload it to AWS Lambda.
