--- elevenlabs-poc/src/lambda.yaml
@@ -1,48 +1,0 @@
-# This AWS SAM template has been generated from your function's configuration. If
-# your function has one or more triggers, note that the AWS resources associated
-# with these triggers aren't fully specified in this template and include
-# placeholder values. Open this template in AWS Infrastructure Composer or your
-# favorite IDE and modify it to specify a serverless application with other AWS
-# resources.
-AWSTemplateFormatVersion: '2010-09-09'
-Transform: AWS::Serverless-2016-10-31
-Description: An AWS Serverless Application Model template describing your function.
-Resources:
-  AdienElevenLabsPOC:
-    Type: AWS::Serverless::Function
-    Properties:
-      CodeUri: .
-      SourceKMSKeyArn: ''
-      Description: ''
-      MemorySize: 128
-      Timeout: 20
-      Handler: main.lambda_handler
-      Runtime: python3.11
-      Architectures:
-        - x86_64
-      EphemeralStorage:
-        Size: 512
-      Environment:
-        Variables:
-          ELEVENLABS_API_KEY: sk_e1cb732d732a91de5627e25c6e30f79953884e306c270324
-      EventInvokeConfig:
-        MaximumEventAgeInSeconds: 21600
-        MaximumRetryAttempts: 2
-      PackageType: Zip
-      Policies:
-        - Statement:
-            - Effect: Allow
-              Action:
-                - logs:PutLogEvents
-                - logs:CreateLogGroup
-                - logs:CreateLogStream
-              Resource: arn:aws:logs:*:*:*
-            - Effect: Allow
-              Action:
-                - s3:PutObject
-              Resource: arn:aws:s3:::adien-apac-practice-bucket/*
-      RecursiveLoop: Terminate
-      SnapStart:
-        ApplyOn: None
-      RuntimeManagementConfig:
-        UpdateRuntimeOn: Auto
