+++ elevenlabs-poc/src/requirements.txt
@@ -1,0 +1,2 @@
+python-dotenv
+elevenlabs
